Finished Problem 1
Write a subprogram that asks the user for the size of a list of integers, then allocates an array of that size, and then repeatedly asks the user for each number and stores it in the next element of the array. Return the pointer to the new array to the calling function (using the 'return' statement).